+++ src/index.ts
@@ -31,6 +31,7 @@ export { default as Progress } from '@/package/components/progress.vue';
 export { default as Badge } from '@/package/components/badge.vue';
 export { default as Avatar } from '@/package/components/avatar.vue';
 export { default as DataTable } from '@/package/components/data-table.vue';
+export { default as Slider } from '@/package/components/slider.vue';
 export type { DataTableColumn } from '@/package/components/data-table.vue';
 // config
 export { default as VueForge, setTheme, updateTheme, getTheme } from '@/package/config/index';