+++ src/package/themes/default/index.ts
@@ -26,6 +26,7 @@ import alert from './components/alert';
 import progress from './components/progress';
 import badge from './components/badge';
 import avatar from './components/avatar';
+import slider from './components/slider';
 
 export default {
     ...base,
@@ -58,5 +59,6 @@ export default {
         progress,
         badge,
         avatar,
+        slider,
     },
 };